requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.12
botocore==1.38.12
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
entrypoints==0.4
Faker==37.1.0
favicon==0.7.0
filelock==3.18.0
folium==0.19.5
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.38.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
prometheus_client==0.21.1
protobuf==5.29.4
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.12.0
setuptools==80.3.1
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.31.1
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.45.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.2.1
streamlit-keyup==0.3.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_folium==0.25.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
wheel==0.45.1
xyzservices==2025.4.0
